app/requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
